+++ src/app/components/DiceCluster.tsx
@@ -1,5 +1,5 @@
-import React from 'react';
-import CustomDice from './CustomDice'; 
+import React from "react";
+import CustomDice from "./CustomDice";
 
 const DiceCluster = () => (
   <div className="relative flex justify-center items-center">
@@ -7,9 +7,9 @@ const DiceCluster = () => (
     <div
       className="absolute"
       style={{
-        transform: 'rotate(-10deg)', 
-        top: '15px',
-        left: '10px',
+        transform: "rotate(-10deg)",
+        top: "15px",
+        left: "10px",
       }}
     >
       <CustomDice label="L" />
@@ -19,8 +19,8 @@ const DiceCluster = () => (
     <div
       className="absolute"
       style={{
-        transform: 'rotate(15deg)', 
-        top: '10px', 
+        transform: "rotate(15deg)",
+        top: "10px",
       }}
     >
       <CustomDice label="C" />
@@ -30,9 +30,9 @@ const DiceCluster = () => (
     <div
       className="absolute"
       style={{
-        transform: 'rotate(-15deg)', 
-        top: '-50px', 
-        left: '5px', 
+        transform: "rotate(-15deg)",
+        top: "-50px",
+        left: "5px",
       }}
     >
       <CustomDice label="R" />
