+++ src/app/fonts.tsx
@@ -1,0 +1,3 @@
+import { Roboto_Mono } from "next/font/google";
+
+export const robotoMono = Roboto_Mono({ subsets: ["latin"] });
