+++ src/app/signup/page.tsx
@@ -1,69 +1,74 @@
-
 "use client";
-import React, { useState } from 'react';
-import '../globals.css';
+
+import React, { useState } from "react";
+import Layout from "@/app/layout";
 
 const SignUpPage: React.FC = () => {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [errorMessage, setErrorMessage] = useState('');
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const [errorMessage, setErrorMessage] = useState("");
 
   const handleSignUp = (e: React.FormEvent) => {
     e.preventDefault();
     //  add logic for user registration
-    if (email === '' || password === '') {
-      setErrorMessage('Please fill in all fields.');
+    if (email === "" || password === "") {
+      setErrorMessage("Please fill in all fields.");
     } else {
-      setErrorMessage('');
+      setErrorMessage("");
       // add  signup logic
     }
   };
 
   return (
-    <div className="min-h-screen flex items-center justify-center bg-gradient-to-r from-black-500 to-indigo-500">
-      <div className="bg-gray-800 p-6 rounded-lg shadow-md w-full sm:w-96">
-        <h2 className="text-2xl font-semibold mb-4 text-white">Sign Up</h2>
+    <Layout>
+      {" "}
+      <div className="min-h-screen flex items-center justify-center bg-gradient-to-r from-black-500 to-indigo-500">
+        <div className="bg-gray-800 p-6 rounded-lg shadow-md w-full sm:w-96">
+          <h2 className="text-2xl font-semibold mb-4 text-white">Sign Up</h2>
 
-        {errorMessage && (
-          <div className="text-red-500 text-sm mb-4">
-            {errorMessage}
-          </div>
-        )}
+          {errorMessage && (
+            <div className="text-red-500 text-sm mb-4">{errorMessage}</div>
+          )}
 
-        <form onSubmit={handleSignUp}>
-          <div className="mb-4">
-            <label htmlFor="email" className="block text-gray-300">Email</label>
-            <input
-              type="email"
-              id="email"
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
-              className="mt-2 w-full px-3 py-2 border border-gray-500 rounded-md bg-gray-700 text-white focus:ring-blue-500 focus:border-blue-500"
-              placeholder="Enter email"
-            />
-          </div>
+          <form onSubmit={handleSignUp}>
+            <div className="mb-4">
+              <label htmlFor="email" className="block text-gray-300">
+                Email
+              </label>
+              <input
+                type="email"
+                id="email"
+                value={email}
+                onChange={(e) => setEmail(e.target.value)}
+                className="mt-2 w-full px-3 py-2 border border-gray-500 rounded-md bg-gray-700 text-white focus:ring-blue-500 focus:border-blue-500"
+                placeholder="Enter email"
+              />
+            </div>
 
-          <div className="mb-4">
-            <label htmlFor="password" className="block text-gray-300">Password</label>
-            <input
-              type="password"
-              id="password"
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
-              className="mt-2 w-full px-3 py-2 border border-gray-500 rounded-md bg-gray-700 text-white focus:ring-blue-500 focus:border-blue-500"
-              placeholder="Enter password"
-            />
-          </div>
+            <div className="mb-4">
+              <label htmlFor="password" className="block text-gray-300">
+                Password
+              </label>
+              <input
+                type="password"
+                id="password"
+                value={password}
+                onChange={(e) => setPassword(e.target.value)}
+                className="mt-2 w-full px-3 py-2 border border-gray-500 rounded-md bg-gray-700 text-white focus:ring-blue-500 focus:border-blue-500"
+                placeholder="Enter password"
+              />
+            </div>
 
-          <button
-            type="submit"
-            className="w-full bg-blue-600 text-white py-2 rounded-md hover:bg-blue-700"
-          >
-            Sign Up
-          </button>
-        </form>
+            <button
+              type="submit"
+              className="w-full bg-blue-600 text-white py-2 rounded-md hover:bg-blue-700"
+            >
+              Sign Up
+            </button>
+          </form>
+        </div>
       </div>
-    </div>
+    </Layout>
   );
 };
