+++ tailwind.config.ts
@@ -12,6 +12,20 @@ export default {
         background: "var(--background)",
         foreground: "var(--foreground)",
       },
+      keyframes: {
+        pulseNotify: {
+          '0%': { transform: 'scale(1)' },          
+          '10%': { transform: 'scale(1.1)' },       
+          '20%': { transform: 'scale(1)' },         
+          '30%': { transform: 'scale(1.1)' },       
+          '40%': { transform: 'scale(1)' },         
+          '80%': { transform: 'scale(1)' },        
+          '100%': { transform: 'scale(1)' },        
+        },
+      },
+      animation: {
+        pulseNotify: 'pulseNotify 4s ease-in-out infinite', 
+      },
     },
   },
   plugins: [],
